feat: Phase 2 - Infrastructure層の実装完了
## 実装内容

### Task 3.1: JIRA統合の分割と移行
- src/infrastructure/external-apis/atlassian/jira/ (6ファイル)
  - types.ts: 型定義
  - client.ts: JIRAClient (HTTP通信、認証、レート制限)
  - adf-converter.ts: Atlassian Document Format変換
  - status-mapper.ts: フェーズラベルマッピング
  - issue-builder.ts: Issue作成ヘルパー
  - sync-service.ts: メイン同期ロジック

### Task 3.2: Confluence統合の分割と移行
- src/infrastructure/external-apis/atlassian/confluence/ (6ファイル)
  - types.ts: ConfluencePage, ConfluenceConfig等
  - client.ts: ConfluenceClient (ページ検索/作成/更新)
  - approval.ts: Markdown変換と承認フロー
  - hierarchy.ts: 階層構造ページ作成
  - sync-service.ts: メイン同期ロジック

### Task 3.3: GitHub API統合の移行
- src/infrastructure/external-apis/github/ (4ファイル)
  - types.ts: Workflow Run、CI Status等の型
  - client.ts: GitHubActionsClient
  - utils.ts: exponentialBackoff, parseGitHubWorkflowRun

### Task 3.4: 設定管理の移行
- src/infrastructure/config/
  - schema.ts: Zodスキーマと型定義
  - default-config.json: デフォルト設定

### Task 3.5: ファイルシステム操作の移行
- src/infrastructure/filesystem/
  - project-meta.ts: プロジェクトメタデータ読み込み

### Task 3.6: パーサーの移行
- src/infrastructure/parsers/
  - markdown-parser.ts: 要件・コンポーネント・フロー抽出
  - aidlc-parser.ts: AI-DLC形式tasks.mdパーサー

### 後方互換性
- scripts/配下の元ファイルは再エクスポート形式に更新
- 既存コードへの影響なし

## テスト結果
- ✅ 型チェック: 全てパス
- ✅ テスト: 818 passed, 5 skipped (823 total)
- ✅ Lint: エラーなし
