+++ src/infrastructure/config/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Configuration Management Module
+ */
+
+// Schema exports
+export * from './schema.js';
