+++ src/infrastructure/external-apis/github/index.ts
@@ -1,0 +1,18 @@
+/**
+ * GitHub API Integration Module
+ */
+
+// Client
+export { GitHubActionsClient } from './client.js';
+
+// Types
+export type {
+  IGitHubWorkflowRun,
+  IRepositoryCIStatus,
+  GitHubAPIError,
+  OctokitError,
+  Result,
+} from './types.js';
+
+// Utils
+export { exponentialBackoff, parseGitHubWorkflowRun } from './utils.js';
