+++ src/infrastructure/filesystem/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Filesystem Module
+ */
+
+// Project metadata
+export type { ProjectMetadata } from './project-meta.js';
+export {
+  loadProjectMeta,
+  formatProjectInfo,
+  getRepositoryInfo,
+} from './project-meta.js';
